+++ src/main/java/lotto/domain/LottoResult.java
@@ -1,4 +1,4 @@
-package lotto.domain.value;
+package lotto.domain;
 
 import java.util.Collections;
 import java.util.EnumMap;
