+++ src/main/java/lotto/domain/LottoTicket.java
@@ -1,4 +1,4 @@
-package lotto.domain.value;
+package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.Collections;
