+++ src/test/java/lotto/domain/LottoResultTest.java
@@ -1,5 +1,8 @@
-package lotto.domain.value;
+package lotto.domain;
 
+import lotto.domain.LottoResult;
+import lotto.domain.Rank;
+import lotto.domain.WinningNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
