+++ src/test/java/lotto/domain/LottoTicketsTest.java
@@ -1,5 +1,7 @@
-package lotto.domain.value;
+package lotto.domain;
 
+import lotto.domain.Rank;
+import lotto.domain.WinningNumbers;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
