+++ src/test/java/lotto/domain/PriceTest.java
@@ -1,5 +1,6 @@
-package lotto.domain.value;
+package lotto.domain;
 
+import lotto.domain.Price;
 import lotto.exception.InvalidUnitException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
