+++ src/test/java/lotto/domain/RankTest.java
@@ -1,5 +1,6 @@
-package lotto.domain.value;
+package lotto.domain;
 
+import lotto.domain.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
